+++ fediseer/consts.py
@@ -1,4 +1,4 @@
-FEDISEER_VERSION = "0.22.1"
+FEDISEER_VERSION = "0.23.0"
 SUPPORTED_SOFTWARE = {
     "lemmy",
     "mastodon",